// src/common/enums/email.enums.ts
export enum EMAIL_EVENTS {
  FORGOT_PASSWORD = 'email.forgot.password',
  VERIFY_EMAIL = 'email.verfiy.email',
  ORDER_CONFIRMED = 'email.order.confirmed',
  ORDER_CANCELLED = 'email.order.cancelled',
  ORDER_REFUNDED = 'email.order.refunded',
  ORDER_STATUS_UPDATED = 'email.order.status_updated',
  USER_BANNED = 'email.user.banned',
  USER_UNBANNED = 'email.user.unbanned',
}